requirements.txt
anyio==3.7.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
fastapi==0.97.0
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.33
numpy==1.24.3
pandas==2.0.2
protobuf==3.20.3
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
tzdata==2023.3
uvicorn==0.22.0
